rag_arxiv_cscl_week4/requirements.txt
arxiv
PyMuPDF
sentence-transformers
faiss-cpu
fastapi
uvicorn
numpy
tqdm